refactor: separate sign-up logic into SignUpClient component for improved readability and maintainability
